agent-audio/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.50.0
google-cloud-texttospeech>=2.17.0